Light adaptation of v0.6.2 (#124)
* Degree units

Adding (lat/lon) degrees as a possible unit.

* Loading the area from the hydro_units

Added a condition to avoid loading the total area of the hydro_units if the cover areas are also given, as then the total area is computed from the cover areas.

* TimeSeries1D - Add timespan selection

* Fixing the SpotpySetup initialization

Fixing the SpotpySetup initialization under the new version of Hydrobricks - change in the handling of the Observations

* Fixing the SpotpySetup initialization - Correction

Fixing the SpotpySetup initialization under the new version of Hydrobricks - change in the handling of the Observations - Correction

* Fixing the function get_total_precipitation()

Fixing the function get_total_precipitation() that did not work after updating Hydrobricks

* Changing the melt_ice aliases for several types of ice

Changing the melt_ice aliases for several types of ice - If there is more than one type of ice (for example clean ice and debris-covered ice), we need several aliases. Currently the code bugs since the aliases for all ice types all contain melt_t_ice.

* Condition to renaming the x and y dimensions

Adds a condition to renaming the x and y dimensions in TimeSeries to avoid a UserWarning if the names were already right.

* Adding a new functionality to create a BehaviourLandCoverChange object

Adding a new functionality to create a BehaviourLandCoverChange object from shapefiles of glacier extents (optionally also from shapefiles of debris-covered glacier extents) from a Catchment object.