Implementing the netcdf, constants
Implementing the creation of the netcdfs by decomposing the input period as Julian days, computing the radiation and bringing back the correct Julian days into a netcdf file. Debugged the transformation of the aspect to radian in the angle of incidence computations. I removed the computation of the hour angle, and exported the between-steps (azimuth, zenith etc as netcdf). I also added the constants.py file that I forgot in the last commit.