Cleaned version
Version that computes the radiation without additional outputs, and print. The radiation obtained presents several very low, negative values, but these values are systematically identical. They are probably incorrectly set to zero. I used the mean height of the DEM to compute the radiation, but I could use the local pixel heights. The formula for the True anomaly is also to be investigated more in details.