Fully glaciated hydro unit cases
Adapting the FixLandCoverFractionsTotal() function to work also with cases where the hydro unit is fully glaciated at the beginning of the Behaviour time serie. Before this fix, the function was returning an error at the beginning of every new simulation since the ground area was left at a positive value from the end of the previous simulation, and the inequality criteria did not allow to set the ground area back to 0.